Rename eventhandler.py file to state.py
