+++ src/isar/state_machine/states/lockdown.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING, List
 
 import isar.state_machine.states.home as Home
 from isar.apis.models.models import LockdownResponse
-from isar.eventhandlers.eventhandler import EventHandlerMapping, State, Transition
+from isar.eventhandlers.state import EventHandlerMapping, State, Transition
 from isar.models.events import EmptyMessage
 from isar.state_machine.states_enum import States
