+++ src/isar/state_machine/states/stopping_paused_return_home.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING, List
 import isar.state_machine.states.monitor as Monitor
 import isar.state_machine.states.return_home_paused as ReturnHomePaused
 from isar.apis.models.models import MissionStartResponse
-from isar.eventhandlers.eventhandler import EventHandlerMapping, State, Transition
+from isar.eventhandlers.state import EventHandlerMapping, State, Transition
 from isar.models.events import AbortedMission, EmptyMessage
 from isar.state_machine.states_enum import States
 from robot_interface.models.mission.mission import Mission
