+++ tests/isar/state_machine/states/test_lockdown_state.py
@@ -1,6 +1,6 @@
 from typing import cast
 
-from isar.eventhandlers.eventhandler import EventHandlerMapping, State
+from isar.eventhandlers.state import EventHandlerMapping, State
 from isar.models.events import EmptyMessage
 from isar.state_machine.state_machine import StateMachine
 from isar.state_machine.states.going_to_lockdown import GoingToLockdown
