+++ tests/isar/state_machine/states/test_maintenance_mode_state.py
@@ -1,6 +1,6 @@
 from typing import cast
 
-from isar.eventhandlers.eventhandler import EventHandlerMapping, State
+from isar.eventhandlers.state import EventHandlerMapping, State
 from isar.state_machine.state_machine import StateMachine
 from isar.state_machine.states.home import Home
 from isar.state_machine.states.maintenance import Maintenance
