+++ tests/isar/state_machine/states/test_monitor_state.py
@@ -9,7 +9,7 @@ from fastapi import HTTPException
 from pytest_mock import MockerFixture
 
 from isar.config.settings import settings
-from isar.eventhandlers.eventhandler import EventHandlerMapping, State
+from isar.eventhandlers.state import EventHandlerMapping, State
 from isar.models.events import EmptyMessage
 from isar.modules import ApplicationContainer
 from isar.services.utilities.scheduling_utilities import SchedulingUtilities
