+++ tests/isar/state_machine/states/test_returning_home_state.py
@@ -1,10 +1,6 @@
 from typing import cast
 
-from isar.eventhandlers.eventhandler import (
-    EventHandlerMapping,
-    State,
-    TimeoutHandlerMapping,
-)
+from isar.eventhandlers.state import EventHandlerMapping, State, TimeoutHandlerMapping
 from isar.models.events import EmptyMessage
 from isar.state_machine.state_machine import StateMachine
 from isar.state_machine.states.await_next_mission import AwaitNextMission
